+++ backend/app/data_products/output_ports/router.py
@@ -25,8 +25,11 @@ from app.core.namespace.validation import (
 from app.data_products.output_ports.curated_queries.router import (
     router as curated_queries_router,
 )
+from app.data_products.output_ports.data_quality.router import (
+    router as data_quality_router,
+)
 from app.data_products.output_ports.model import Dataset as OutputPortModel
-from app.data_products.output_ports.model import ensure_dataset_exists
+from app.data_products.output_ports.model import ensure_output_port_exists
 from app.data_products.output_ports.query_stats.router import (
     router as query_stats_router,
 )
@@ -102,6 +105,7 @@ old_route = "/datasets"
 route = "/v2/data_products/{data_product_id}/output_ports"
 router.include_router(query_stats_router)
 router.include_router(curated_queries_router)
+router.include_router(data_quality_router)
 
 
 @router.get(f"{old_route}/namespace_suggestion", deprecated=True)
@@ -193,7 +197,7 @@ def get_output_port(
 def get_event_history_old(
     id: UUID, db: Session = Depends(get_db_session)
 ) -> Sequence[GetEventHistoryResponseItemOld]:
-    ds = ensure_dataset_exists(id, db)
+    ds = ensure_output_port_exists(id, db)
     return EventService(db).get_history(ds.id, EventReferenceEntity.DATASET)
 
 
@@ -201,7 +205,7 @@ def get_event_history_old(
 def get_output_ports_event_history(
     data_product_id: UUID, id: UUID, db: Session = Depends(get_db_session)
 ) -> GetEventHistoryResponse:
-    ds = ensure_dataset_exists(id, db, data_product_id=data_product_id)
+    ds = ensure_output_port_exists(id, db, data_product_id=data_product_id)
     return GetEventHistoryResponse(
         events=[
             GetEventHistoryResponseItemOld.model_validate(ds).convert()
@@ -305,7 +309,7 @@ def remove_dataset_old(
     db: Session = Depends(get_db_session),
     authenticated_user: User = Depends(get_authenticated_user),
 ) -> None:
-    ds = ensure_dataset_exists(id, db)
+    ds = ensure_output_port_exists(id, db)
     return remove_dataset(ds.data_product_id, id, db, authenticated_user)
 
 
@@ -370,7 +374,7 @@ def update_dataset(
     db: Session = Depends(get_db_session),
     authenticated_user: User = Depends(get_authenticated_user),
 ) -> UpdateOutputPortResponse:
-    ds = ensure_dataset_exists(id, db)
+    ds = ensure_output_port_exists(id, db)
     return update_output_port(ds.data_product_id, id, dataset, db, authenticated_user)
 
 
@@ -434,7 +438,7 @@ def update_dataset_about(
     db: Session = Depends(get_db_session),
     authenticated_user: User = Depends(get_authenticated_user),
 ) -> None:
-    ds = ensure_dataset_exists(id, db)
+    ds = ensure_output_port_exists(id, db)
     return update_output_port_about(
         ds.data_product_id, id, dataset, db, authenticated_user
     )
@@ -496,7 +500,7 @@ def update_dataset_status(
     db: Session = Depends(get_db_session),
     authenticated_user: User = Depends(get_authenticated_user),
 ) -> None:
-    ds = ensure_dataset_exists(id, db)
+    ds = ensure_output_port_exists(id, db)
     return update_output_port_status(
         ds.data_product_id, id, dataset, db, authenticated_user
     )
@@ -565,7 +569,7 @@ def update_dataset_usage(
 def get_graph_data_old(
     id: UUID, db: Session = Depends(get_db_session), level: int = 3
 ) -> Graph:
-    ds = ensure_dataset_exists(id, db)
+    ds = ensure_output_port_exists(id, db)
     return get_output_port_graph_data(ds.data_product_id, id, db, level)
 
 
@@ -595,7 +599,7 @@ def set_value_for_dataset(
     db: Session = Depends(get_db_session),
     authenticated_user: User = Depends(get_authenticated_user),
 ) -> None:
-    ds = ensure_dataset_exists(id, db)
+    ds = ensure_output_port_exists(id, db)
     return set_value_for_output_port(
         ds.data_product_id, id, setting_id, value, db, authenticated_user
     )
@@ -617,7 +621,7 @@ def set_value_for_output_port(
     db: Session = Depends(get_db_session),
     authenticated_user: User = Depends(get_authenticated_user),
 ) -> None:
-    ensure_dataset_exists(id, db, data_product_id=data_product_id)
+    ensure_output_port_exists(id, db, data_product_id=data_product_id)
     DataProductSettingService(db).set_value_for_product(setting_id, id, value)
     EventService(db).create_event(
         CreateEvent(